Moved types of app components to separate files
