+++ code/src/components/app-box/AppBox.types.ts
@@ -1,0 +1,4 @@
+import { HTMLProps } from "react";
+import { BoxProps } from "@mui/material/Box";
+
+export type AppBoxProps<TElement> = BoxProps & HTMLProps<TElement>;
