+++ code/src/components/app-button/AppButton.types.ts
@@ -1,0 +1,17 @@
+import { ButtonProps } from "@mui/material/Button";
+
+type ButtonVariant =
+  | "contained"
+  | "outlined"
+  | "text"
+  | "danger"
+  | "light"
+  | "dark"
+  | "shadow";
+type ButtonSize = "small" | "medium" | "large" | "extra-large";
+
+export type AppButtonProps = Omit<ButtonProps, "variant" | "size"> & {
+  variant?: ButtonVariant;
+  size?: ButtonSize;
+  isLoading?: boolean;
+};
