+++ code/src/components/app-container/AppContainer.types.ts
@@ -1,0 +1,3 @@
+import { ContainerProps } from "@mui/material/Container";
+
+export type AppContainerProps = ContainerProps;
