+++ code/src/components/app-logo/AppLogo.types.ts
@@ -1,0 +1,8 @@
+import { ComponentPropsWithoutRef } from "react";
+
+export const COMPONENT = "img";
+
+export type AppLogoType = Omit<
+  ComponentPropsWithoutRef<typeof COMPONENT>,
+  "component" | "src" | "alt"
+>;
